DASH-723 Move ManagePageCOntainer components into own folders
