+++ client/src/components/ManagePageContainer/components/DateComponent/DateComponent.module.scss
@@ -1,5 +1,5 @@
-@import '../../../assets/styles/mediaQueriesMixings.scss';
-@import '../../../assets/styles/colors.scss';
+@import '../../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../../assets/styles/colors.scss';
 
 .warnIcon {
   margin-left: 10px;
