+++ client/src/components/ManagePageContainer/components/DateComponent/index.ts
@@ -1,0 +1,1 @@
+export * from './DateComponent';
