+++ client/src/components/ManagePageContainer/components/FioAddress/FioAddress.module.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/styles/colors.scss';
+@import '../../../../assets/styles/colors.scss';
 
 .nameContainer {
   display: flex;
