+++ client/src/components/ManagePageContainer/components/FioAddress/index.ts
@@ -1,0 +1,1 @@
+export * from './FioAddress';
