+++ client/src/components/ManagePageContainer/components/ItemCopmonent/ItemComponent.module.scss
@@ -1,5 +1,5 @@
-@import '../../../assets/styles/colors.scss';
-@import '../../../assets/styles/fontMixins.scss';
+@import '../../../../assets/styles/colors.scss';
+@import '../../../../assets/styles/fontMixins.scss';
 
 .title {
   font-size: 1.25em;
