+++ client/src/components/ManagePageContainer/components/ItemCopmonent/index.ts
@@ -1,0 +1,1 @@
+export * from './ItemComponent';
