+++ client/src/components/ManagePageContainer/components/ItemsScreenView/index.ts
@@ -1,0 +1,2 @@
+export { DesktopView } from './DesktopView';
+export { MobileView } from './MobileView';
