+++ client/src/components/ManagePageContainer/components/SettingsItem/SettingsItem.module.scss
@@ -1,6 +1,6 @@
-@import '../../../assets/styles/colors.scss';
-@import '../../../assets/styles/fontMixins.scss';
-@import '../../../assets/styles/mediaQueriesMixings.scss';
+@import '../../../../assets/styles/colors.scss';
+@import '../../../../assets/styles/fontMixins.scss';
+@import '../../../../assets/styles/mediaQueriesMixings.scss';
 
 .settingsContainer {
   width: 100%;
