+++ client/src/components/ManagePageContainer/components/SettingsItem/index.ts
@@ -1,0 +1,1 @@
+export * from './SettingsItem';
